Mongo：编辑 Int64 不再窄化成 Int32（driver 仅对 Int64 输出 $numberLong）
relaxed extjson 下 Int32/Int64 都是裸数字无法区分，编辑 Int64 正小值会经 serde 反序列化
窄成 Int32（静默改类型）。driver 改为只对 Int64 特化包装 $numberLong（其余仍 relaxed、
显示不变），cell 标独立 long kind、编辑按 $numberLong 还原；flatten 已把 extjson 包装
当标量故不误判嵌套。补 5 个测试锁定往返与 Int32 不受影响。

+++ crates/ramag-tool-mongodb/src/views/result_panel/cell.rs
@@ -72,11 +72,11 @@ pub(super) fn extjson_cell(map: &serde_json::Map<String, Value>) -> Option<Cell>
             kind: "binary",
         });
     }
-    // Int64
+    // Int64（driver 特化包装成 $numberLong，与 Int32 分开标 "long"，编辑时按 64 位还原）
     if let Some(v) = map.get("$numberLong").and_then(|x| x.as_str()) {
         return Some(Cell {
             text: v.to_string(),
-            kind: "int",
+            kind: "long",
         });
     }
     // Int32 canonical
